chore(outputs): add common methods to finding class
